requirements.txt
numpy
opencv-python
matplotlib
tqdm
scikit-learn
fastapi
uvicorn[standard]
pydantic
onnx
onnxruntime
pillow
python-multipart
